aes_web_app/requirements.txt
Flask==2.0.1
pycryptodome==3.10.1